fix minor bugs

+++ Back/src/main/java/com/aberdote/OVPN4ALL/dto/LogDTO.java
@@ -8,7 +8,7 @@ import lombok.Data;
 @Data @Builder @AllArgsConstructor
 public class LogDTO {
 
-    private Integer lineNumber;
+    private Long lineNumber;
     private String content;
 
 }
